requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
finnhub-python==2.4.19
Flask==3.0.1
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
polygon-api-client==1.13.4
python-dateutil==2.8.2
requests==2.31.0
setuptools==69.0.3
six==1.16.0
urllib3==1.26.18
websockets==12.0
Werkzeug==3.0.1
